+++ files/data/scripts/chromium_settings.sh
@@ -1,0 +1,6 @@
+# settings for /scripts/start_chromium.sh
+
+URL="https://192.168.1.144:8123/lovelace/"
+SCALE="1.0"
+EXTRA_CHROMIUM_ARGS=""
+EXTRA_XORG_ARGS="-nocursor"