+++ scripts/start_buttons.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-# start the buttons monitoring service
-
-export DISPLAY=:0
-/usr/bin/python3 /scripts/buttons_app.py